refactor: modularize dedup system
