+++ src/core/dedup/__init__.py
@@ -1,0 +1,4 @@
+"""Deduplication package providing detection, remediation and reporting."""
+from .facade import DeduplicationFacade
+
+__all__ = ["DeduplicationFacade"]
