+++ src/core/dedup/algorithms/__init__.py
@@ -1,0 +1,5 @@
+"""Algorithm implementations for deduplication."""
+from .detector import DuplicationDetector
+from .parser import CodeParser
+
+__all__ = ["DuplicationDetector", "CodeParser"]
