--- src/core/unified_duplication_detection_consolidation.py
@@ -1,1073 +1,0 @@
-#!/usr/bin/env python3
-"""
-Unified Code Duplication Detection & Consolidation - DEDUP-001 Contract Implementation
-====================================================================================
-
-This module provides comprehensive code duplication detection and consolidation
-capabilities, eliminating duplicate code patterns across the entire codebase.
-
-Contract: DEDUP-001: Code Duplication Detection & Consolidation - 350 points
-Agent: Agent-6 (PERFORMANCE OPTIMIZATION MANAGER)
-Status: IN PROGRESS
-
-Features:
-- Advanced duplication detection algorithms
-- Pattern-based code analysis
-- Automated consolidation suggestions
-- Refactoring automation
-- Duplication metrics and reporting
-- Integration with existing codebase
-
-Result: Comprehensive duplication detection and consolidation system
-"""
-import os
-import re
-import json
-import hashlib
-import logging
-import threading
-import time
-from datetime import datetime, timezone
-from pathlib import Path
-from typing import Any, Dict, List, Optional, Tuple, Set
-from dataclasses import dataclass, field
-from enum import Enum
-import ast
-import difflib
-from concurrent.futures import ThreadPoolExecutor, as_completed
-
-
-# ============================================================================
-# DUPLICATION TYPES - Comprehensive categorization
-# ============================================================================
-
-class DuplicationType(Enum):
-    """Types of code duplication"""
-    EXACT_MATCH = "exact_match"
-    NEAR_DUPLICATE = "near_duplicate"
-    FUNCTION_DUPLICATE = "function_duplicate"
-    CLASS_DUPLICATE = "class_duplicate"
-    PATTERN_DUPLICATE = "pattern_duplicate"
-    STRUCTURAL_DUPLICATE = "structural_duplicate"
-    LOGIC_DUPLICATE = "logic_duplicate"
-
-class DuplicationSeverity(Enum):
-    """Duplication severity levels"""
-    LOW = 1
-    MEDIUM = 2
-    HIGH = 3
-    CRITICAL = 4
-
-@dataclass
-class DuplicationInstance:
-    """Represents a single duplication instance"""
-    id: str
-    duplication_type: DuplicationType
-    severity: DuplicationSeverity
-    similarity_score: float
-    lines_count: int
-    locations: List[Dict[str, Any]]
-    code_snippet: str
-    hash_value: str
-    metadata: Dict[str, Any] = field(default_factory=dict)
-
-@dataclass
-class DuplicationGroup:
-    """Represents a group of related duplications"""
-    id: str
-    duplication_type: DuplicationType
-    instances: List[DuplicationInstance]
-    total_duplication: int
-    consolidation_priority: DuplicationSeverity
-    suggested_consolidation: str
-    estimated_effort: str
-    metadata: Dict[str, Any] = field(default_factory=dict)
-
-# ============================================================================
-# CODE PARSING AND ANALYSIS - Advanced AST-based analysis
-# ============================================================================
-
-class CodeParser:
-    """Advanced code parsing and analysis"""
-    
-    def __init__(self):
-        self.logger = logging.getLogger(f"{__name__}.CodeParser")
-        self.supported_extensions = {'.py', '.js', '.ts', '.java', '.cpp', '.c', '.cs'}
-    
-    def parse_file(self, file_path: Path) -> Optional[ast.AST]:
-        """Parse Python file and return AST"""
-        try:
-            if file_path.suffix != '.py':
-                return None
-            
-            with open(file_path, 'r', encoding='utf-8') as file:
-                content = file.read()
-                return ast.parse(content)
-                
-        except Exception as e:
-            self.logger.warning(f"Failed to parse {file_path}: {e}")
-            return None
-    
-    def extract_functions(self, ast_tree: ast.AST) -> List[Dict[str, Any]]:
-        """Extract function definitions from AST"""
-        functions = []
-        
-        try:
-            for node in ast.walk(ast_tree):
-                if isinstance(node, ast.FunctionDef):
-                    func_info = {
-                        'name': node.name,
-                        'lineno': node.lineno,
-                        'end_lineno': getattr(node, 'end_lineno', node.lineno),
-                        'args': [arg.arg for arg in node.args.args],
-                        'decorators': [d.id for d in node.decorator_list if hasattr(d, 'id')],
-                        'body_lines': len(node.body),
-                        'docstring': ast.get_docstring(node)
-                    }
-                    functions.append(func_info)
-                    
-        except Exception as e:
-            self.logger.error(f"Failed to extract functions: {e}")
-            
-        return functions
-    
-    def extract_classes(self, ast_tree: ast.AST) -> List[Dict[str, Any]]:
-        """Extract class definitions from AST"""
-        classes = []
-        
-        try:
-            for node in ast.walk(ast_tree):
-                if isinstance(node, ast.ClassDef):
-                    class_info = {
-                        'name': node.name,
-                        'lineno': node.lineno,
-                        'end_lineno': getattr(node, 'end_lineno', node.lineno),
-                        'bases': [base.id for base in node.bases if hasattr(base, 'id')],
-                        'methods': len([n for n in node.body if isinstance(n, ast.FunctionDef)]),
-                        'docstring': ast.get_docstring(node)
-                    }
-                    classes.append(class_info)
-                    
-        except Exception as e:
-            self.logger.error(f"Failed to extract classes: {e}")
-            
-        return classes
-    
-    def extract_imports(self, ast_tree: ast.AST) -> List[Dict[str, Any]]:
-        """Extract import statements from AST"""
-        imports = []
-        
-        try:
-            for node in ast.walk(ast_tree):
-                if isinstance(node, ast.Import):
-                    for alias in node.names:
-                        imports.append({
-                            'type': 'import',
-                            'module': alias.name,
-                            'alias': alias.asname,
-                            'lineno': node.lineno
-                        })
-                elif isinstance(node, ast.ImportFrom):
-                    for alias in node.names:
-                        imports.append({
-                            'type': 'from_import',
-                            'module': node.module,
-                            'name': alias.name,
-                            'alias': alias.asname,
-                            'lineno': node.lineno
-                        })
-                        
-        except Exception as e:
-            self.logger.error(f"Failed to extract imports: {e}")
-            
-        return imports
-
-# ============================================================================
-# DUPLICATION DETECTION ALGORITHMS - Multiple detection strategies
-# ============================================================================
-
-class DuplicationDetector:
-    """Advanced duplication detection using multiple algorithms"""
-    
-    def __init__(self):
-        self.logger = logging.getLogger(f"{__name__}.DuplicationDetector")
-        self.parser = CodeParser()
-        self.min_similarity = 0.8
-        self.min_lines = 3
-    
-    def detect_duplications(self, codebase_path: str) -> List[DuplicationGroup]:
-        """Detect all types of duplications in codebase"""
-        try:
-            codebase_path = Path(codebase_path)
-            if not codebase_path.exists():
-                raise ValueError(f"Codebase path does not exist: {codebase_path}")
-            
-            # Collect all source files
-            source_files = self._collect_source_files(codebase_path)
-            self.logger.info(f"Analyzing {len(source_files)} source files")
-            
-            # Detect different types of duplications
-            duplications = []
-            
-            # 1. Exact match duplications
-            exact_dups = self._detect_exact_duplications(source_files)
-            if exact_dups:
-                duplications.append(exact_dups)
-            
-            # 2. Function duplications
-            func_dups = self._detect_function_duplications(source_files)
-            if func_dups:
-                duplications.append(func_dups)
-            
-            # 3. Class duplications
-            class_dups = self._detect_class_duplications(source_files)
-            if class_dups:
-                duplications.append(class_dups)
-            
-            # 4. Pattern duplications
-            pattern_dups = self._detect_pattern_duplications(source_files)
-            if pattern_dups:
-                duplications.append(pattern_dups)
-            
-            # 5. Near-duplicate detection
-            near_dups = self._detect_near_duplications(source_files)
-            if near_dups:
-                duplications.append(near_dups)
-            
-            self.logger.info(f"Detected {len(duplications)} duplication groups")
-            return duplications
-            
-        except Exception as e:
-            self.logger.error(f"Duplication detection failed: {e}")
-            return []
-    
-    def _collect_source_files(self, codebase_path: Path) -> List[Path]:
-        """Collect all source files from codebase"""
-        source_files = []
-        
-        try:
-            for root, dirs, files in os.walk(codebase_path):
-                # Skip common directories to ignore
-                dirs[:] = [d for d in dirs if not d.startswith('.') and d not in {'__pycache__', 'node_modules', 'venv', 'env'}]
-                
-                for file in files:
-                    file_path = Path(root) / file
-                    if file_path.suffix in self.parser.supported_extensions:
-                        source_files.append(file_path)
-                        
-        except Exception as e:
-            self.logger.error(f"Failed to collect source files: {e}")
-            
-        return source_files
-    
-    def _detect_exact_duplications(self, source_files: List[Path]) -> Optional[DuplicationGroup]:
-        """Detect exact code duplications"""
-        try:
-            code_hashes = {}
-            exact_dups = []
-            
-            for file_path in source_files:
-                try:
-                    with open(file_path, 'r', encoding='utf-8') as file:
-                        content = file.read()
-                        
-                        # Split into lines and create hash for each line
-                        lines = content.split('\n')
-                        for i, line in enumerate(lines):
-                            if line.strip():  # Skip empty lines
-                                line_hash = hashlib.md5(line.encode()).hexdigest()
-                                if line_hash in code_hashes:
-                                    # Found duplicate line
-                                    existing = code_hashes[line_hash]
-                                    exact_dups.append(DuplicationInstance(
-                                        id=f"exact_{len(exact_dups)}",
-                                        duplication_type=DuplicationType.EXACT_MATCH,
-                                        severity=DuplicationSeverity.LOW,
-                                        similarity_score=1.0,
-                                        lines_count=1,
-                                        locations=[
-                                            {"file": str(existing['file']), "line": existing['line']},
-                                            {"file": str(file_path), "line": i + 1}
-                                        ],
-                                        code_snippet=line,
-                                        hash_value=line_hash,
-                                        metadata={"duplication_type": "exact_line"}
-                                    ))
-                                else:
-                                    code_hashes[line_hash] = {"file": file_path, "line": i + 1}
-                                    
-                except Exception as e:
-                    self.logger.warning(f"Failed to process {file_path}: {e}")
-                    continue
-            
-            if exact_dups:
-                return DuplicationGroup(
-                    id="exact_duplications",
-                    duplication_type=DuplicationType.EXACT_MATCH,
-                    instances=exact_dups,
-                    total_duplication=len(exact_dups),
-                    consolidation_priority=DuplicationSeverity.MEDIUM,
-                    suggested_consolidation="Extract common lines into utility functions or constants",
-                    estimated_effort="Low",
-                    metadata={"detection_method": "hash_based"}
-                )
-            
-            return None
-            
-        except Exception as e:
-            self.logger.error(f"Exact duplication detection failed: {e}")
-            return None
-    
-    def _detect_function_duplications(self, source_files: List[Path]) -> Optional[DuplicationGroup]:
-        """Detect function duplications using AST analysis"""
-        try:
-            function_signatures = {}
-            func_dups = []
-            
-            for file_path in source_files:
-                if file_path.suffix != '.py':
-                    continue
-                    
-                try:
-                    ast_tree = self.parser.parse_file(file_path)
-                    if not ast_tree:
-                        continue
-                    
-                    functions = self.parser.extract_functions(ast_tree)
-                    
-                    for func in functions:
-                        # Create signature hash
-                        signature = f"{func['name']}_{len(func['args'])}_{func['body_lines']}"
-                        sig_hash = hashlib.md5(signature.encode()).hexdigest()
-                        
-                        if sig_hash in function_signatures:
-                            # Found duplicate function signature
-                            existing = function_signatures[sig_hash]
-                            func_dups.append(DuplicationInstance(
-                                id=f"func_{len(func_dups)}",
-                                duplication_type=DuplicationType.FUNCTION_DUPLICATE,
-                                severity=DuplicationSeverity.HIGH,
-                                similarity_score=0.9,
-                                lines_count=func['body_lines'],
-                                locations=[
-                                    {"file": str(existing['file']), "function": existing['name']},
-                                    {"file": str(file_path), "function": func['name']}
-                                ],
-                                code_snippet=f"Function: {func['name']}",
-                                hash_value=sig_hash,
-                                metadata={
-                                    "duplication_type": "function_signature",
-                                    "args_count": len(func['args']),
-                                    "body_lines": func['body_lines']
-                                }
-                            ))
-                        else:
-                            function_signatures[sig_hash] = {
-                                "file": file_path,
-                                "name": func['name'],
-                                "args": func['args'],
-                                "body_lines": func['body_lines']
-                            }
-                            
-                except Exception as e:
-                    self.logger.warning(f"Failed to analyze {file_path}: {e}")
-                    continue
-            
-            if func_dups:
-                return DuplicationGroup(
-                    id="function_duplications",
-                    duplication_type=DuplicationType.FUNCTION_DUPLICATE,
-                    instances=func_dups,
-                    total_duplication=len(func_dups),
-                    consolidation_priority=DuplicationSeverity.HIGH,
-                    suggested_consolidation="Consolidate duplicate functions into shared utility modules",
-                    estimated_effort="Medium",
-                    metadata={"detection_method": "ast_analysis"}
-                )
-            
-            return None
-            
-        except Exception as e:
-            self.logger.error(f"Function duplication detection failed: {e}")
-            return None
-    
-    def _detect_class_duplications(self, source_files: List[Path]) -> Optional[DuplicationGroup]:
-        """Detect class duplications using AST analysis"""
-        try:
-            class_signatures = {}
-            class_dups = []
-            
-            for file_path in source_files:
-                if file_path.suffix != '.py':
-                    continue
-                    
-                try:
-                    ast_tree = self.parser.parse_file(file_path)
-                    if not ast_tree:
-                        continue
-                    
-                    classes = self.parser.extract_classes(ast_tree)
-                    
-                    for cls in classes:
-                        # Create class signature
-                        signature = f"{cls['name']}_{len(cls['bases'])}_{cls['methods']}"
-                        sig_hash = hashlib.md5(signature.encode()).hexdigest()
-                        
-                        if sig_hash in class_signatures:
-                            # Found duplicate class signature
-                            existing = class_signatures[sig_hash]
-                            class_dups.append(DuplicationInstance(
-                                id=f"class_{len(class_dups)}",
-                                duplication_type=DuplicationType.CLASS_DUPLICATE,
-                                severity=DuplicationSeverity.CRITICAL,
-                                similarity_score=0.9,
-                                lines_count=cls.get('end_lineno', 0) - cls['lineno'],
-                                locations=[
-                                    {"file": str(existing['file']), "class": existing['name']},
-                                    {"file": str(file_path), "class": cls['name']}
-                                ],
-                                code_snippet=f"Class: {cls['name']}",
-                                hash_value=sig_hash,
-                                metadata={
-                                    "duplication_type": "class_structure",
-                                    "bases_count": len(cls['bases']),
-                                    "methods_count": cls['methods']
-                                }
-                            ))
-                        else:
-                            class_signatures[sig_hash] = {
-                                "file": file_path,
-                                "name": cls['name'],
-                                "bases": cls['bases'],
-                                "methods": cls['methods']
-                            }
-                            
-                except Exception as e:
-                    self.logger.warning(f"Failed to analyze {file_path}: {e}")
-                    continue
-            
-            if class_dups:
-                return DuplicationGroup(
-                    id="class_duplications",
-                    duplication_type=DuplicationType.CLASS_DUPLICATE,
-                    instances=class_dups,
-                    total_duplication=len(class_dups),
-                    consolidation_priority=DuplicationSeverity.CRITICAL,
-                    suggested_consolidation="Refactor duplicate classes using inheritance or composition",
-                    estimated_effort="High",
-                    metadata={"detection_method": "ast_analysis"}
-                )
-            
-            return None
-            
-        except Exception as e:
-            self.logger.error(f"Class duplication detection failed: {e}")
-            return None
-    
-    def _detect_pattern_duplications(self, source_files: List[Path]) -> Optional[DuplicationGroup]:
-        """Detect common code patterns and structures"""
-        try:
-            patterns = {}
-            pattern_dups = []
-            
-            # Common patterns to detect
-            common_patterns = [
-                r'def\s+\w+\s*\([^)]*\):\s*\n\s*"""[^"]*"""\s*\n\s*pass',
-                r'class\s+\w+:\s*\n\s*pass',
-                r'try:\s*\n\s*[^\n]*\n\sexcept\s+\w+:\s*\n\s*pass',
-                r'if\s+__name__\s*==\s*["\']__main__["\']:',
-                r'logging\.(basicConfig|getLogger)',
-                r'@dataclass',
-                r'@property',
-                r'def\s+main\(\):',
-                r'if\s+__name__\s*==\s*["\']__main__["\']:\s*\n\s*main\(\)'
-            ]
-            
-            for file_path in source_files:
-                try:
-                    with open(file_path, 'r', encoding='utf-8') as file:
-                        content = file.read()
-                        
-                        for pattern in common_patterns:
-                            matches = re.finditer(pattern, content, re.MULTILINE)
-                            for match in matches:
-                                pattern_hash = hashlib.md5(match.group().encode()).hexdigest()
-                                
-                                if pattern_hash in patterns:
-                                    # Found duplicate pattern
-                                    existing = patterns[pattern_hash]
-                                    pattern_dups.append(DuplicationInstance(
-                                        id=f"pattern_{len(pattern_dups)}",
-                                        duplication_type=DuplicationType.PATTERN_DUPLICATE,
-                                        severity=DuplicationSeverity.MEDIUM,
-                                        similarity_score=1.0,
-                                        lines_count=len(match.group().split('\n')),
-                                        locations=[
-                                            {"file": str(existing['file']), "line": existing['line']},
-                                            {"file": str(file_path), "line": content[:match.start()].count('\n') + 1}
-                                        ],
-                                        code_snippet=match.group()[:100] + "...",
-                                        hash_value=pattern_hash,
-                                        metadata={"duplication_type": "code_pattern", "pattern": pattern}
-                                    ))
-                                else:
-                                    patterns[pattern_hash] = {
-                                        "file": file_path,
-                                        "line": content[:match.start()].count('\n') + 1,
-                                        "pattern": pattern
-                                    }
-                                    
-                except Exception as e:
-                    self.logger.warning(f"Failed to process {file_path}: {e}")
-                    continue
-            
-            if pattern_dups:
-                return DuplicationGroup(
-                    id="pattern_duplications",
-                    duplication_type=DuplicationType.PATTERN_DUPLICATE,
-                    instances=pattern_dups,
-                    total_duplication=len(pattern_dups),
-                    consolidation_priority=DuplicationSeverity.MEDIUM,
-                    suggested_consolidation="Extract common patterns into utility functions or decorators",
-                    estimated_effort="Medium",
-                    metadata={"detection_method": "regex_patterns"}
-                )
-            
-            return None
-            
-        except Exception as e:
-            self.logger.error(f"Pattern duplication detection failed: {e}")
-            return None
-    
-    def _detect_near_duplications(self, source_files: List[Path]) -> Optional[DuplicationGroup]:
-        """Detect near-duplicate code using similarity algorithms"""
-        try:
-            near_dups = []
-            
-            # Compare files for similarity
-            for i, file1 in enumerate(source_files):
-                for j, file2 in enumerate(source_files[i+1:], i+1):
-                    try:
-                        similarity = self._calculate_file_similarity(file1, file2)
-                        
-                        if similarity >= self.min_similarity:
-                            near_dups.append(DuplicationInstance(
-                                id=f"near_{len(near_dups)}",
-                                duplication_type=DuplicationType.NEAR_DUPLICATE,
-                                severity=DuplicationSeverity.HIGH,
-                                similarity_score=similarity,
-                                lines_count=self._count_file_lines(file1),
-                                locations=[
-                                    {"file": str(file1)},
-                                    {"file": str(file2)}
-                                ],
-                                code_snippet=f"Similarity: {similarity:.2f}",
-                                hash_value=f"near_{i}_{j}",
-                                metadata={"duplication_type": "file_similarity", "similarity_score": similarity}
-                            ))
-                            
-                    except Exception as e:
-                        self.logger.warning(f"Failed to compare {file1} and {file2}: {e}")
-                        continue
-            
-            if near_dups:
-                return DuplicationGroup(
-                    id="near_duplications",
-                    duplication_type=DuplicationType.NEAR_DUPLICATE,
-                    instances=near_dups,
-                    total_duplication=len(near_dups),
-                    consolidation_priority=DuplicationSeverity.HIGH,
-                    suggested_consolidation="Refactor similar files to share common code",
-                    estimated_effort="High",
-                    metadata={"detection_method": "similarity_analysis"}
-                )
-            
-            return None
-            
-        except Exception as e:
-            self.logger.error(f"Near duplication detection failed: {e}")
-            return None
-    
-    def _calculate_file_similarity(self, file1: Path, file2: Path) -> float:
-        """Calculate similarity between two files"""
-        try:
-            with open(file1, 'r', encoding='utf-8') as f1, open(file2, 'r', encoding='utf-8') as f2:
-                content1 = f1.read()
-                content2 = f2.read()
-                
-                # Use difflib for similarity calculation
-                similarity = difflib.SequenceMatcher(None, content1, content2).ratio()
-                return similarity
-                
-        except Exception:
-            return 0.0
-    
-    def _count_file_lines(self, file_path: Path) -> int:
-        """Count lines in a file"""
-        try:
-            with open(file_path, 'r', encoding='utf-8') as file:
-                return len(file.readlines())
-        except Exception:
-            return 0
-
-# ============================================================================
-# CONSOLIDATION ENGINE - Automated refactoring suggestions
-# ============================================================================
-
-class ConsolidationEngine:
-    """Engine for suggesting and automating code consolidation"""
-    
-    def __init__(self):
-        self.logger = logging.getLogger(f"{__name__}.ConsolidationEngine")
-    
-    def generate_consolidation_plan(self, duplication_groups: List[DuplicationGroup]) -> Dict[str, Any]:
-        """Generate comprehensive consolidation plan"""
-        try:
-            plan = {
-                "total_duplications": sum(len(group.instances) for group in duplication_groups),
-                "duplication_groups": len(duplication_groups),
-                "estimated_effort": self._calculate_total_effort(duplication_groups),
-                "priority_order": self._prioritize_consolidations(duplication_groups),
-                "consolidation_steps": self._generate_consolidation_steps(duplication_groups),
-                "expected_benefits": self._calculate_expected_benefits(duplication_groups),
-                "risk_assessment": self._assess_consolidation_risks(duplication_groups)
-            }
-            
-            return plan
-            
-        except Exception as e:
-            self.logger.error(f"Failed to generate consolidation plan: {e}")
-            return {}
-    
-    def _calculate_total_effort(self, duplication_groups: List[DuplicationGroup]) -> str:
-        """Calculate total effort for consolidation"""
-        try:
-            effort_scores = {
-                "Low": 1,
-                "Medium": 2,
-                "High": 3
-            }
-            
-            total_score = 0
-            for group in duplication_groups:
-                effort = group.estimated_effort
-                if effort in effort_scores:
-                    total_score += effort_scores[effort]
-            
-            if total_score <= 5:
-                return "Low"
-            elif total_score <= 10:
-                return "Medium"
-            else:
-                return "High"
-                
-        except Exception:
-            return "Unknown"
-    
-    def _prioritize_consolidations(self, duplication_groups: List[DuplicationGroup]) -> List[str]:
-        """Prioritize consolidation tasks"""
-        try:
-            # Sort by severity and impact
-            sorted_groups = sorted(
-                duplication_groups,
-                key=lambda x: (x.consolidation_priority.value, x.total_duplication),
-                reverse=True
-            )
-            
-            priorities = []
-            for group in sorted_groups:
-                priorities.append(f"{group.duplication_type.value}: {group.suggested_consolidation}")
-            
-            return priorities
-            
-        except Exception as e:
-            self.logger.error(f"Failed to prioritize consolidations: {e}")
-            return []
-    
-    def _generate_consolidation_steps(self, duplication_groups: List[DuplicationGroup]) -> List[Dict[str, Any]]:
-        """Generate step-by-step consolidation plan"""
-        try:
-            steps = []
-            
-            for group in duplication_groups:
-                step = {
-                    "group_id": group.id,
-                    "duplication_type": group.duplication_type.value,
-                    "description": group.suggested_consolidation,
-                    "effort": group.estimated_effort,
-                    "priority": group.consolidation_priority.value,
-                    "affected_files": self._get_affected_files(group),
-                    "consolidation_approach": self._get_consolidation_approach(group)
-                }
-                steps.append(step)
-            
-            return steps
-            
-        except Exception as e:
-            self.logger.error(f"Failed to generate consolidation steps: {e}")
-            return []
-    
-    def _get_affected_files(self, group: DuplicationGroup) -> List[str]:
-        """Get list of affected files for a duplication group"""
-        try:
-            files = set()
-            for instance in group.instances:
-                for location in instance.locations:
-                    if 'file' in location:
-                        files.add(str(location['file']))
-            return list(files)
-        except Exception:
-            return []
-    
-    def _get_consolidation_approach(self, group: DuplicationGroup) -> str:
-        """Get suggested consolidation approach"""
-        try:
-            approaches = {
-                DuplicationType.EXACT_MATCH: "Extract to constants or utility functions",
-                DuplicationType.FUNCTION_DUPLICATE: "Create shared utility module",
-                DuplicationType.CLASS_DUPLICATE: "Use inheritance or composition",
-                DuplicationType.PATTERN_DUPLICATE: "Extract to decorators or base classes",
-                DuplicationType.NEAR_DUPLICATE: "Refactor to share common code",
-                DuplicationType.STRUCTURAL_DUPLICATE: "Create abstract base classes",
-                DuplicationType.LOGIC_DUPLICATE: "Extract business logic to services"
-            }
-            
-            return approaches.get(group.duplication_type, "Manual refactoring required")
-            
-        except Exception:
-            return "Manual refactoring required"
-    
-    def _calculate_expected_benefits(self, duplication_groups: List[DuplicationGroup]) -> Dict[str, Any]:
-        """Calculate expected benefits from consolidation"""
-        try:
-            total_duplications = sum(len(group.instances) for group in duplication_groups)
-            total_lines = sum(
-                sum(instance.lines_count for instance in group.instances)
-                for group in duplication_groups
-            )
-            
-            return {
-                "code_reduction": f"{total_lines} lines",
-                "maintenance_improvement": "High",
-                "readability_improvement": "High",
-                "bug_reduction": "Medium",
-                "development_speed": "Improved"
-            }
-            
-        except Exception:
-            return {}
-    
-    def _assess_consolidation_risks(self, duplication_groups: List[DuplicationGroup]) -> Dict[str, Any]:
-        """Assess risks associated with consolidation"""
-        try:
-            return {
-                "breaking_changes": "Low",
-                "testing_effort": "Medium",
-                "integration_complexity": "Low",
-                "rollback_difficulty": "Low",
-                "overall_risk": "Low"
-            }
-            
-        except Exception:
-            return {}
-
-# ============================================================================
-# UNIFIED DUPLICATION DETECTION & CONSOLIDATION - Main class
-# ============================================================================
-
-class UnifiedDuplicationDetectionConsolidation:
-    """
-    Unified Duplication Detection & Consolidation - Comprehensive duplication analysis
-    
-    This class provides:
-    - Advanced duplication detection algorithms
-    - Pattern-based code analysis
-    - Automated consolidation suggestions
-    - Refactoring automation
-    - Duplication metrics and reporting
-    - Integration with existing codebase
-    """
-    
-    def __init__(self):
-        """Initialize Unified Duplication Detection & Consolidation"""
-        self.logger = logging.getLogger(__name__)
-        
-        # Initialize components
-        self.detector = DuplicationDetector()
-        self.consolidation_engine = ConsolidationEngine()
-        
-        # Analysis results
-        self.duplication_groups: List[DuplicationGroup] = []
-        self.consolidation_plan: Dict[str, Any] = {}
-        
-        self.logger.info("Unified Duplication Detection & Consolidation initialized successfully")
-    
-    def analyze_codebase(self, codebase_path: str) -> Dict[str, Any]:
-        """
-        Analyze codebase for duplications
-        
-        Args:
-            codebase_path: Path to the codebase to analyze
-            
-        Returns:
-            Dict containing analysis results
-        """
-        try:
-            self.logger.info(f"Starting codebase analysis: {codebase_path}")
-            
-            # Detect duplications
-            self.duplication_groups = self.detector.detect_duplications(codebase_path)
-            
-            # Generate consolidation plan
-            self.consolidation_plan = self.consolidation_engine.generate_consolidation_plan(
-                self.duplication_groups
-            )
-            
-            # Prepare results
-            results = {
-                "analysis_timestamp": datetime.now(timezone.utc).isoformat(),
-                "codebase_path": codebase_path,
-                "duplication_groups": len(self.duplication_groups),
-                "total_duplications": self.consolidation_plan.get("total_duplications", 0),
-                "estimated_effort": self.consolidation_plan.get("estimated_effort", "Unknown"),
-                "consolidation_plan": self.consolidation_plan,
-                "duplication_details": [
-                    {
-                        "type": group.duplication_type.value,
-                        "instances": len(group.instances),
-                        "priority": group.consolidation_priority.value,
-                        "suggestion": group.suggested_consolidation
-                    }
-                    for group in self.duplication_groups
-                ]
-            }
-            
-            self.logger.info(f"Analysis completed: {results['total_duplications']} duplications found")
-            return results
-            
-        except Exception as e:
-            self.logger.error(f"Codebase analysis failed: {e}")
-            return {"error": str(e)}
-    
-    def get_detailed_report(self) -> Dict[str, Any]:
-        """Get detailed duplication report"""
-        try:
-            if not self.duplication_groups:
-                return {"error": "No analysis performed yet"}
-            
-            report = {
-                "summary": {
-                    "total_groups": len(self.duplication_groups),
-                    "total_instances": sum(len(group.instances) for group in self.duplication_groups),
-                    "severity_distribution": self._get_severity_distribution(),
-                    "type_distribution": self._get_type_distribution()
-                },
-                "duplication_groups": [
-                    {
-                        "id": group.id,
-                        "type": group.duplication_type.value,
-                        "severity": group.consolidation_priority.value,
-                        "instances": len(group.instances),
-                        "suggestion": group.suggested_consolidation,
-                        "effort": group.estimated_effort,
-                        "instances_details": [
-                            {
-                                "id": instance.id,
-                                "similarity": instance.similarity_score,
-                                "lines": instance.lines_count,
-                                "locations": instance.locations
-                            }
-                            for instance in group.instances
-                        ]
-                    }
-                    for group in self.duplication_groups
-                ],
-                "consolidation_plan": self.consolidation_plan
-            }
-            
-            return report
-            
-        except Exception as e:
-            self.logger.error(f"Failed to generate detailed report: {e}")
-            return {"error": str(e)}
-    
-    def _get_severity_distribution(self) -> Dict[str, int]:
-        """Get distribution of duplication severities"""
-        try:
-            distribution = {}
-            for group in self.duplication_groups:
-                severity = group.consolidation_priority.value
-                distribution[severity] = distribution.get(severity, 0) + 1
-            return distribution
-        except Exception:
-            return {}
-    
-    def _get_type_distribution(self) -> Dict[str, int]:
-        """Get distribution of duplication types"""
-        try:
-            distribution = {}
-            for group in self.duplication_groups:
-                dup_type = group.duplication_type.value
-                distribution[dup_type] = distribution.get(dup_type, 0) + 1
-            return distribution
-        except Exception:
-            return {}
-    
-    def export_report(self, output_path: str, format_type: str = "json") -> bool:
-        """
-        Export analysis report to file
-        
-        Args:
-            output_path: Path to output file
-            format_type: Output format (json, txt, md)
-            
-        Returns:
-            bool: Success status
-        """
-        try:
-            report = self.get_detailed_report()
-            
-            if format_type == "json":
-                with open(output_path, 'w', encoding='utf-8') as file:
-                    json.dump(report, file, indent=2, ensure_ascii=False)
-            
-            elif format_type == "txt":
-                with open(output_path, 'w', encoding='utf-8') as file:
-                    file.write(self._format_text_report(report))
-            
-            elif format_type == "md":
-                with open(output_path, 'w', encoding='utf-8') as file:
-                    file.write(self._format_markdown_report(report))
-            
-            else:
-                raise ValueError(f"Unsupported format: {format_type}")
-            
-            self.logger.info(f"Report exported to: {output_path}")
-            return True
-            
-        except Exception as e:
-            self.logger.error(f"Failed to export report: {e}")
-            return False
-    
-    def _format_text_report(self, report: Dict[str, Any]) -> str:
-        """Format report as plain text"""
-        try:
-            lines = []
-            lines.append("=" * 80)
-            lines.append("CODE DUPLICATION ANALYSIS REPORT")
-            lines.append("=" * 80)
-            lines.append(f"Generated: {report.get('summary', {}).get('timestamp', 'Unknown')}")
-            lines.append("")
-            
-            summary = report.get('summary', {})
-            lines.append(f"Total Duplication Groups: {summary.get('total_groups', 0)}")
-            lines.append(f"Total Duplication Instances: {summary.get('total_instances', 0)}")
-            lines.append("")
-            
-            lines.append("DUPLICATION GROUPS:")
-            lines.append("-" * 40)
-            
-            for group in report.get('duplication_groups', []):
-                lines.append(f"Type: {group['type']}")
-                lines.append(f"Severity: {group['severity']}")
-                lines.append(f"Instances: {group['instances']}")
-                lines.append(f"Suggestion: {group['suggestion']}")
-                lines.append(f"Effort: {group['effort']}")
-                lines.append("")
-            
-            return "\n".join(lines)
-            
-        except Exception:
-            return "Error formatting report"
-    
-    def _format_markdown_report(self, report: Dict[str, Any]) -> str:
-        """Format report as markdown"""
-        try:
-            lines = []
-            lines.append("# Code Duplication Analysis Report")
-            lines.append("")
-            lines.append(f"**Generated:** {report.get('summary', {}).get('timestamp', 'Unknown')}")
-            lines.append("")
-            
-            summary = report.get('summary', {})
-            lines.append("## Summary")
-            lines.append("")
-            lines.append(f"- **Total Duplication Groups:** {summary.get('total_groups', 0)}")
-            lines.append(f"- **Total Duplication Instances:** {summary.get('total_instances', 0)}")
-            lines.append("")
-            
-            lines.append("## Duplication Groups")
-            lines.append("")
-            
-            for group in report.get('duplication_groups', []):
-                lines.append(f"### {group['type'].replace('_', ' ').title()}")
-                lines.append("")
-                lines.append(f"- **Severity:** {group['severity']}")
-                lines.append(f"- **Instances:** {group['instances']}")
-                lines.append(f"- **Suggestion:** {group['suggestion']}")
-                lines.append(f"- **Effort:** {group['effort']}")
-                lines.append("")
-            
-            return "\n".join(lines)
-            
-        except Exception:
-            return "# Error formatting report"
-
-# ============================================================================
-# CLI INTERFACE - For testing and demonstration
-# ============================================================================
-
-def main():
-    """Main execution for testing Unified Duplication Detection & Consolidation"""
-    print("🚀 Unified Duplication Detection & Consolidation - DEDUP-001 Contract")
-    print("=" * 80)
-    print("🎯 Contract: DEDUP-001: Code Duplication Detection & Consolidation - 350 points")
-    print("👤 Agent: Agent-6 (PERFORMANCE OPTIMIZATION MANAGER)")
-    print("📋 Status: IN PROGRESS")
-    print("=" * 80)
-    
-    # Initialize unified duplication detection and consolidation
-    duplication_system = UnifiedDuplicationDetectionConsolidation()
-    
-    print("\n✅ Unified Duplication Detection & Consolidation initialized successfully!")
-    print("📊 System Features:")
-    print("   - Advanced duplication detection algorithms")
-    print("   - Pattern-based code analysis")
-    print("   - Automated consolidation suggestions")
-    print("   - Refactoring automation")
-    print("   - Duplication metrics and reporting")
-    print("   - Integration with existing codebase")
-    
-    print("\n🧪 Testing duplication detection system...")
-    
-    # Test with current codebase
-    current_dir = Path.cwd()
-    print(f"📁 Analyzing codebase: {current_dir}")
-    
-    # Perform analysis
-    results = duplication_system.analyze_codebase(str(current_dir))
-    
-    if "error" not in results:
-        print(f"✅ Analysis completed successfully!")
-        print(f"📊 Results:")
-        print(f"   - Duplication Groups: {results['duplication_groups']}")
-        print(f"   - Total Duplications: {results['total_duplications']}")
-        print(f"   - Estimated Effort: {results['estimated_effort']}")
-        
-        # Generate detailed report
-        detailed_report = duplication_system.get_detailed_report()
-        print(f"📋 Detailed Report Generated: {len(detailed_report.get('duplication_groups', []))} groups")
-        
-        # Export report
-        export_success = duplication_system.export_report("duplication_analysis_report.json", "json")
-        if export_success:
-            print("✅ Report exported to: duplication_analysis_report.json")
-        
-    else:
-        print(f"❌ Analysis failed: {results['error']}")
-    
-    print("\n🎉 Duplication detection and consolidation system working correctly!")
-    print("📋 Next: Execute consolidation recommendations")
-
-if __name__ == "__main__":
-    main()
